+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `olmo_core.generate.chat` module to allow interacting with OlmoCore models without conversion to other formats.
 - Added `GAPMonitorCallback` for monitoring gradients, activations, and parameters (GAP).
 - Added official OLMo3-7B and -32B pretraining scripts and data mix.
-- Added official OLMo3-7B midtraining script and data mix.
+- Added official OLMo3-7B and -32B midtraining script and data mix.
 - Added official OLMo3-7B long-context script and data mix.
 - Added a `NoOpOptimizer` that does nothing, uses no memory, and can be used for debugging.
 - Added official config for Olmo 3 32B.